backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
python-multipart==0.0.9
pydantic-settings==2.5.2
mediapipe==0.10.14
pillow==10.4.0
numpy==1.26.4
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
